az-jenkins-controller: deploy oauth2-proxy for jenkins
This configures caddy to delegate authentication to a deployment of
oauth2-proxy, which will redirect the user to GitHub for authentication.

We peek at the user profile, and in case they're part of the tiiuae
organization, let them in.

As GitHub redirects back to the application after the users consent, we
need to have one GitHub application per Jenkins deployment.
Said GitHub OAuth application also needs to have access to see
membership in the tiiuae application.

This will become easier in the future by adding a "trampoline" IdP
like Dex in between, allowing us to manage individual clients and their
redirect URIs locally, with only the Dex URL needing to be configured
in GitHub.

Once the user gets redirected back, it's passed through to Jenkins.
We inject certain X-Forwarded-* headers into the request, which Jenkins
picks up.

Signed-off-by: Florian Klink <flokli@flokli.de>