az-jenkins-controller: use TOML for casc
This changes the jenkins-casc variable to be a data structure, which can
be extended later on.

We write it out to JSON (which is also valid YAML of course), to be
re-ingested by Jenkins.

We cannot keep YAML as an input format, as Nix cannot parse YAML without
using IFD.

Signed-off-by: Florian Klink <flokli@flokli.de>